make Months

Month statistics

makeMonths(dailyBootOut, eList)

Arguments

dailyBootOut
data frame returned from genDailyBoot

eList
is the data with a fitted model already done. Note that the eList$Sample may have multiple values on a given day and it can also have censored values.

Value

a list of 2 data frame with monthSeq column that corresponds to the months in the "MonthSeq" column in the Daily data frame. The remaining columns are quantiles of the flux or concentration (depending on the data frame).

Examples

eList <- EGRET::Choptank_eList
dailyBoot <- genDailyBoot(eList, nBoot = 2, nKalman = 2)

#> Boot:  1 

#> 
#> There are 216 duplicated Sample dates.

#> 
#>  made surfaces from boot sample 1 replicate     - Kalman index:  1 

#> This function is currently in development

#> % complete:
#> 0 	1 	2 	3 	4 	5 	6 	7 	8 	9 	10 	
#> 11 	12 	13 	14 	15 	16 	17 	18 	19 	20 	
#> 21 	22 	23 	24 	25 	26 	27 	28 	29 	30 	
#> 31 	32 	33 	34 	35 	36 	37 	38 	39 	40 	
#> 41 	42 	43 	44 	45 	46 	47 	48 	49 	50 	
#> 51 	52 	53 	54 	55 	56 	57 	58 	59 	60 	
#> 61 	62 	63 	64 	65 	66 	67 	68 	69 	70 	
#> 71 	72 	73 	74 	75 	76 	77 	78 	79 	80 	
#> 81 	82 	83 	84 	85 	86 	87 	88 	89 	90 	
#> 91 	92 	93 	94 	95 	96 	97 	98 	99 	     - Kalman index:  2 

#> This function is currently in development

#> % complete:
#> 0 	1 	2 	3 	4 	5 	6 	7 	8 	9 	10 	
#> 11 	12 	13 	14 	15 	16 	17 	18 	19 	20 	
#> 21 	22 	23 	24 	25 	26 	27 	28 	29 	30 	
#> 31 	32 	33 	34 	35 	36 	37 	38 	39 	40 	
#> 41 	42 	43 	44 	45 	46 	47 	48 	49 	50 	
#> 51 	52 	53 	54 	55 	56 	57 	58 	59 	60 	
#> 61 	62 	63 	64 	65 	66 	67 	68 	69 	70 	
#> 71 	72 	73 	74 	75 	76 	77 	78 	79 	80 	
#> 81 	82 	83 	84 	85 	86 	87 	88 	89 	90 	
#> 91 	92 	93 	94 	95 	96 	97 	98 	99 	Boot:  2 

#> 
#> There are 222 duplicated Sample dates.

#> 
#>  made surfaces from boot sample 2 replicate     - Kalman index:  1 

#> This function is currently in development

#> % complete:
#> 0 	1 	2 	3 	4 	5 	6 	7 	8 	9 	10 	
#> 11 	12 	13 	14 	15 	16 	17 	18 	19 	20 	
#> 21 	22 	23 	24 	25 	26 	27 	28 	29 	30 	
#> 31 	32 	33 	34 	35 	36 	37 	38 	39 	40 	
#> 41 	42 	43 	44 	45 	46 	47 	48 	49 	50 	
#> 51 	52 	53 	54 	55 	56 	57 	58 	59 	60 	
#> 61 	62 	63 	64 	65 	66 	67 	68 	69 	70 	
#> 71 	72 	73 	74 	75 	76 	77 	78 	79 	80 	
#> 81 	82 	83 	84 	85 	86 	87 	88 	89 	90 	
#> 91 	92 	93 	94 	95 	96 	97 	98 	99 	     - Kalman index:  2 

#> This function is currently in development

#> % complete:
#> 0 	1 	2 	3 	4 	5 	6 	7 	8 	9 	10 	
#> 11 	12 	13 	14 	15 	16 	17 	18 	19 	20 	
#> 21 	22 	23 	24 	25 	26 	27 	28 	29 	30 	
#> 31 	32 	33 	34 	35 	36 	37 	38 	39 	40 	
#> 41 	42 	43 	44 	45 	46 	47 	48 	49 	50 	
#> 51 	52 	53 	54 	55 	56 	57 	58 	59 	60 	
#> 61 	62 	63 	64 	65 	66 	67 	68 	69 	70 	
#> 71 	72 	73 	74 	75 	76 	77 	78 	79 	80 	
#> 81 	82 	83 	84 	85 	86 	87 	88 	89 	90 	
#> 91 	92 	93 	94 	95 	96 	97 	98 	99 

monthPcts <- makeMonths(dailyBoot, eList)
head(monthPcts[["month_flux_results"]])

#>   monthSeq       p1     p2.5       p5      p10      p25      p50      p75
#> 1     1558 220.2434 220.2434 220.2434 220.2434 220.4511 221.9800 222.9231
#> 2     1559 333.2451 333.2451 333.2451 333.2451 333.7003 336.0990 341.0414
#> 3     1560 361.1960 361.1960 361.1960 361.1960 361.7323 364.1975 365.3309
#> 4     1561 504.8057 504.8057 504.8057 504.8057 507.3254 515.2750 515.7581
#> 5     1562 280.9274 280.9274 280.9274 280.9274 281.2043 282.8741 283.9574
#> 6     1563 586.5801 586.5801 586.5801 586.5801 587.1761 589.7077 594.8150
#>        p90      p95    p97.5      p99
#> 1 222.9355 222.9355 222.9355 222.9355
#> 2 342.3445 342.3445 342.3445 342.3445
#> 3 365.4233 365.4233 365.4233 365.4233
#> 4 515.7891 515.7891 515.7891 515.7891
#> 5 284.0388 284.0388 284.0388 284.0388
#> 6 596.2695 596.2695 596.2695 596.2695

head(monthPcts[["month_conc_results"]])

#>   monthSeq        p1      p2.5        p5       p10       p25       p50
#> 1     1558 0.6082637 0.6082637 0.6082637 0.6082637 0.6082832 0.6109218
#> 2     1559 0.7913560 0.7913560 0.7913560 0.7913560 0.7922857 0.7978833
#> 3     1560 1.2022527 1.2022527 1.2022527 1.2022527 1.2038002 1.2114930
#> 4     1561 1.0232673 1.0232673 1.0232673 1.0232673 1.0279356 1.0443349
#> 5     1562 1.1613039 1.1613039 1.1613039 1.1613039 1.1628240 1.1706372
#> 6     1563 0.9901511 0.9901511 0.9901511 0.9901511 0.9913247 0.9952281
#>         p75       p90       p95     p97.5       p99
#> 1 0.6141030 0.6143033 0.6143033 0.6143033 0.6143033
#> 2 0.8101712 0.8133310 0.8133310 0.8133310 0.8133310
#> 3 1.2154828 1.2157959 1.2157959 1.2157959 1.2157959
#> 4 1.0467411 1.0467450 1.0467450 1.0467450 1.0467450
#> 5 1.1751547 1.1755762 1.1755762 1.1755762 1.1755762
#> 6 1.0014897 1.0034493 1.0034493 1.0034493 1.0034493