The bootstrap uncertainty analysis for runGroups results

This function that does the uncertainty analysis for determining the change between two groups of years. The process is virtually identical to what is used for runPairsBoot.

runGroupsBoot(
  eList,
  groupResults,
  nBoot = 100,
  startSeed = 494817,
  blockLength = 200
)

Arguments

eList
named list with at least the Daily, Sample, and INFO dataframes

groupResults
data frame returned from runGroups

nBoot
the maximum number of bootstrap replicates to be used, typically 100

startSeed
setSeed value. Defaults to 494817. This is used to make repeatable output.

blockLength
days, typically 200 is a good choice

Value

eBoot, a named list with bootOut,wordsOut,xConc,xFlux values. bootOut is a data frame with the results of the bootstrapping tests. wordsOut is a character vector describing the results. xConc, xFlux are vectors of length iBoot, of the change in flow normalized concentration or flux computed by each bootstrap replicate (mg/L). pConc and pFlux are vectors of length iBoot, of the change in flow normalized concentration or flux computed from each bootstrap replicate expressed as

See also

Examples

library(EGRET)
eList <- Choptank_eList

if (FALSE) {
groupResults <- runGroups(eList,
                          group1firstYear = 1995,
                          group1lastYear = 2004,
                          group2firstYear = 2005,
                          group2lastYear = 2014,
                          windowSide = 7, wall = TRUE,
                          sample1EndDate = "2004-10-30",
                          paStart = 4, paLong = 2,
                          verbose = FALSE)

boot_group_out <- runGroupsBoot(eList, groupResults)

plotHistogramTrend(eList, boot_group_out, caseSetUp=NA)
}