Allows user to set window parameters for the WRTDS model prior to running the bootstrap procedure

Adds window parameters to INFO file in eList.

setForBoot(eList, caseSetUp, windowY = 7, windowQ = 2, windowS = 0.5,
  edgeAdjust = TRUE)

Arguments

eList

named list with at least the Daily, Sample, and INFO dataframes. Created from the EGRET package, after running modelEstimation.

caseSetUp

data frame returned from trendSetUp.

windowY

numeric specifying the half-window width in the time dimension, in units of years, default is 7.

windowQ

numeric specifying the half-window width in the discharge dimension, units are natural log units, default is 2.

windowS

numeric specifying the half-window with in the seasonal dimension, in units of years, default is 0.5.

edgeAdjust

logical specifying whether to use the modified method for calculating the windows at the edge of the record, default is TRUE.

Value

eList list with Daily,Sample, INFO data frames and surface matrix.

Examples

eList <- EGRET::Choptank_eList

caseSetUp <- trendSetUp(eList,
  year1=1985, 
  year2=2005,
  nBoot = 50, 
  bootBreak = 39,
  blockLength = 200)
#> Sample set runs from 1979 to 2011
#> year1 = 1985 this is the first water year of trend period
#> year2 = 2005 this is the last water year of trend period
#> nBoot = 50 this is the maximum number of replicates that will be run
#> bootBreak = 39 this is the minimum number of replicates that will be run
#> blockLength = 200 this is the number of days in a bootstrap block

bootSetUp <- setForBoot(eList,caseSetUp)