make Months

Month statistics using WRTDSKalman bootstrapping approach. The input to this function is the dailyBootOut matrix which contains nReplicate sets of daily flux values for the period of interest. The results are in the form of quantiles of concentration and of flux for each of these months.

makeMonths(dailyBootOut, eList)

Arguments

dailyBootOut
data frame returned from genDailyBoot

eList
is the data with a fitted model already done. Note that the eList$Sample may have multiple values on a given day and it can also have censored values.

Value

a list of 2 data frames, one for average concentration, in mg/L and one for flux (in kg/day) In each data frame the first column is monthSeq that corresponds to the months in the "MonthSeq" column in the eList$Daily data frame. The remaining columns are quantiles of the flux or concentration (depending on the data frame).

Examples

eList <- EGRET::Choptank_eList
dailyBoot <- genDailyBoot(eList, nBoot = 2, nKalman = 2)

#> Boot:  1 

#> 
#> There are 246 duplicated Sample dates.

#> 
#>  made surfaces from boot sample 1 replicate     - Kalman index:  1 
#> % complete:
#> 100 	
#>      - Kalman index:  2 
#> % complete:
#> 100 	
#> Boot:  2 

#> 
#> There are 205 duplicated Sample dates.

#> 
#>  made surfaces from boot sample 2 replicate     - Kalman index:  1 
#> % complete:
#> 100 	
#>      - Kalman index:  2 
#> % complete:
#> 100 

monthPcts <- makeMonths(dailyBoot, eList)
head(monthPcts[["month_flux_results"]])

#>   monthSeq       p1     p2.5       p5      p10      p25      p50      p75
#> 1     1558 199.6863 199.6863 199.6863 199.6863 201.5338 217.2171 229.7662
#> 2     1559 270.3779 270.3779 270.3779 270.3779 282.4721 324.8338 345.6275
#> 3     1560 319.8392 319.8392 319.8392 319.8392 323.0918 334.0781 360.2775
#> 4     1561 455.8086 455.8086 455.8086 455.8086 459.6790 478.9452 491.4650
#> 5     1562 263.3969 263.3969 263.3969 263.3969 266.2830 281.5678 306.8437
#> 6     1563 508.5761 508.5761 508.5761 508.5761 530.8875 602.3057 631.3287
#>        p90      p95    p97.5      p99
#> 1 230.5690 230.5690 230.5690 230.5690
#> 2 350.5324 350.5324 350.5324 350.5324
#> 3 368.6010 368.6010 368.6010 368.6010
#> 4 493.0865 493.0865 493.0865 493.0865
#> 5 313.0602 313.0602 313.0602 313.0602
#> 6 639.5084 639.5084 639.5084 639.5084

head(monthPcts[["month_conc_results"]])

#>   monthSeq        p1      p2.5        p5       p10       p25       p50
#> 1     1558 0.5691069 0.5691069 0.5691069 0.5691069 0.5716131 0.5961196
#> 2     1559 0.6515980 0.6515980 0.6515980 0.6515980 0.6774443 0.7735794
#> 3     1560 1.0774534 1.0774534 1.0774534 1.0774534 1.0877790 1.1199318
#> 4     1561 0.9136589 0.9136589 0.9136589 0.9136589 0.9247895 0.9585959
#> 5     1562 1.0904925 1.0904925 1.0904925 1.0904925 1.1021427 1.1700176
#> 6     1563 0.8628881 0.8628881 0.8628881 0.8628881 0.8911019 0.9864602
#>         p75       p90       p95     p97.5       p99
#> 1 0.6295418 0.6350199 0.6350199 0.6350199 0.6350199
#> 2 0.8212985 0.8310061 0.8310061 0.8310061 0.8310061
#> 3 1.1922823 1.2160071 1.2160071 1.2160071 1.2160071
#> 4 1.0063340 1.0221086 1.0221086 1.0221086 1.0221086
#> 5 1.2802168 1.3059751 1.3059751 1.3059751 1.3059751
#> 6 1.0298333 1.0407187 1.0407187 1.0407187 1.0407187